organized the components structure
moves components from inside the page to outside the pages folder.
