+++ components/layout/Navbar.jsx
@@ -34,8 +34,6 @@ const NavBar = () => {
 
                   <NavDropdown title="Programmes" id="basic-nav-dropdown">
                     <NavDropdown.Item href="/programmes">Programmes</NavDropdown.Item>
-
-                    <NavDropdown.Item href="/programmes">Beela Programees</NavDropdown.Item>
                     <NavDropdown.Item href="/pollination-for-mentees">
                       Polination for Mentees
                     </NavDropdown.Item>
