+++ pages/about.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from "next"
 import Head from "next/head"
-import Navbar from "../components/Navbar"
-import Footer from "../components/Footer"
+import Navbar from "../components/layout/Navbar"
+import Footer from "../components/layout/Footer"
 import avatar from "../public/images/avatar.png"
 import Image from "next/image"
-import AboutBeela from "./components/about/AboutBeela"
+import AboutBeela from "../components/sections/AboutBeela"
 
 const About: NextPage = () => {
   return (
