+++ pages/index.tsx
@@ -2,9 +2,9 @@ import type { NextPage } from "next"
 import Popper from "@popperjs/core"
 import Link from "next/link"
 import Head from "next/head"
-import Navbar from "../components/Navbar"
-import Footer from "../components/Footer"
-import Testimonials from "../components/Testimonials"
+import Navbar from "../components/layout/Navbar"
+import Footer from "../components/layout/Footer"
+import Testimonials from "../components/sections/Testimonials"
 import { Hero } from "../components/Hero"
 import { TeamCard } from "../components/TeamCard"
 // import avatar from "../public/avatar.png"
