+++ pages/pollination-for-mentees.tsx
@@ -2,9 +2,9 @@ import type { NextPage } from "next"
 import Popper from "@popperjs/core"
 import Link from "next/link"
 import Head from "next/head"
-import Navbar from "../components/Navbar"
-import Footer from "../components/Footer"
-import Testimonials from "../components/Testimonials"
+import Navbar from "../components/layout/Navbar"
+import Footer from "../components/layout/Footer"
+import Testimonials from "../components/sections/Testimonials"
 import React, { Component } from "react"
 import "react-responsive-carousel/lib/styles/carousel.min.css"
 import { Carousel } from "react-responsive-carousel"
